requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
colorama==0.4.3
Django==3.0.6
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-otp==0.9.1
djangorestframework==3.11.0
idna==2.9
PyJWT==1.7.1
pytz==2020.1
qrcode==6.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
twilio==6.40.0
urllib3==1.25.9
Pillow
whitenoise==5.0.1
gunicorn==20.0.4
django-rest-passwordreset